+++ src/main/kotlin/com/cinephilia/payment/domain/Payment.kt
@@ -1,6 +1,6 @@
-package com.cinephiliapaymentservice.domain
+package com.cinephilia.payment.domain
 
-import com.cinephiliapaymentservice.domain.enums.PaymentStatus
+import com.cinephilia.payment.domain.enums.PaymentStatus
 import ru.quipy.core.annotations.AggregateType
 import ru.quipy.domain.Aggregate
 import ru.quipy.domain.AggregateState
