+++ src/main/kotlin/com/cinephilia/payment/domain/enums/PaymentStatus.kt
@@ -1,4 +1,4 @@
-package com.cinephiliapaymentservice.domain.enums
+package com.cinephilia.payment.domain.enums
 
 enum class PaymentStatus {
     New,
